requirements-dev.txt
black==23.7.0
isort==5.12.0
pytest-timeout==2.1.0
requests==2.31.0
